+++ rust/signer/src/string.rs
@@ -43,3 +43,19 @@ impl StringPtr {
 	}
 }
 
+// string ffi
+
+#[no_mangle]
+pub unsafe extern fn rust_string_ptr(s: *mut String) -> *mut StringPtr {
+  Box::into_raw(Box::new(StringPtr::from(&**s)))
+}
+
+#[no_mangle]
+pub unsafe extern fn rust_string_destroy(s: *mut String) {
+  let _ = Box::from_raw(s);
+}
+
+#[no_mangle]
+pub unsafe extern fn rust_string_ptr_destroy(s: *mut StringPtr) {
+  let _ = Box::from_raw(s);
+}
